+++ sdk/communication/Azure.Communication.Identity/tests/CommunicationIdentityClient/CommunicationIdentityClientLiveTests.cs
@@ -22,11 +22,21 @@ namespace Azure.Communication.Identity.Tests
     public class CommunicationIdentityClientLiveTests : CommunicationIdentityClientLiveTestBase
     {
         /// <summary>
+        /// Options used to exchange an AAD access token of a Teams user for a new Communication Identity access token.
+        /// </summary>
+        private GetTokenForTeamsUserOptions CTEOptions = new GetTokenForTeamsUserOptions("Sanitized", "Sanitized", "Sanitized");
+        /// <summary>
         /// Initializes a new instance of the <see cref="CommunicationIdentityClient"/> class.
         /// </summary>
         /// <param name="isAsync">A flag used by the Azure Core Test Framework to differentiate between tests for asynchronous and synchronous methods.</param>
         public CommunicationIdentityClientLiveTests(bool isAsync) : base(isAsync)
         {
+        }
+
+        [OneTimeSetUp]
+        public async Task InitVariables()
+        {
+            CTEOptions = await CreateTeamsUserParams();
         }
 
         [Test]
@@ -131,61 +141,50 @@ namespace Azure.Communication.Identity.Tests
         }
 
         [Test]
-        public async Task GetTokenForTeamsUserWithValidToken()
+        public async Task GetTokenForTeamsUserWithValidParameters()
         {
             if (TestEnvironment.ShouldIgnoreIdentityExchangeTokenTest) {
                 Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
             }
 
-            string token = await generateTeamsToken();
-
             CommunicationIdentityClient client = CreateClientWithConnectionString();
-            Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(token);
+            Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(CTEOptions);
             Assert.IsNotNull(tokenResponse.Value);
             Assert.IsFalse(string.IsNullOrWhiteSpace(tokenResponse.Value.Token));
         }
 
         [Test]
-        public async Task GetTokenForTeamsUserWithEmptyTokenShouldThrow()
+        [TestCase(true, false, false, "token", TestName = "GetTokenForTeamsUserWithNullTokenShouldThrow")]
+        [TestCase(false, true, false, "appId", TestName = "GetTokenForTeamsUserWithNullAppIdShouldThrow")]
+        [TestCase(false, false, true, "userId", TestName = "GetTokenForTeamsUserWithNullUserIdShouldThrow")]
+        [TestCase(true, true, true, "token", TestName = "GetTokenForTeamsUserWithNullParamsShouldThrow")]
+        public async Task GetTokenForTeamsUserWithNullParamsShouldThrow(bool isTokenNull, bool isAppIdNull, bool isUserIdNull, string paramName)
         {
+            var teamsUserAadToken = isTokenNull ? null : CTEOptions.TeamsUserAadToken;
+            var clientId = isAppIdNull ? null : CTEOptions.ClientId;
+            var userObjectId = isUserIdNull ? null : CTEOptions.UserObjectId;
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync("");
-            }
-            catch (RequestFailedException ex)
-            {
-                Assert.NotNull(ex.Message);
-                Assert.True(ex.Message.Contains("401"));
-                Console.WriteLine(ex.Message);
-                return;
-            }
-            Assert.Fail("An exception should have been thrown.");
-        }
-
-        [Test]
-        public async Task GetTokenForTeamsUserWithNullTokenShouldThrow()
-        {
-            try
-            {
-                CommunicationIdentityClient client = CreateClientWithConnectionString();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(null);
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(teamsUserAadToken, clientId, userObjectId));
             }
             catch (ArgumentNullException ex)
             {
-                Assert.AreEqual("token", ex.ParamName);
+                Assert.AreEqual(paramName, ex.ParamName);
                 return;
             }
             Assert.Fail("An exception should have been thrown.");
         }
 
         [Test]
-        public async Task GetTokenForTeamsUserWithInvalidTokenShouldThrow()
+        [TestCase("", TestName = "GetTokenForTeamsUserWithEmptyTokenShouldThrow")]
+        [TestCase("invalid", TestName = "GetTokenForTeamsUserWithInvalidTokenShouldThrow")]
+        public async Task GetTokenForTeamsUserWithInvalidTokenShouldThrow(string token)
         {
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync("invalid");
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(token, CTEOptions.ClientId, CTEOptions.UserObjectId));
             }
             catch (RequestFailedException ex)
             {
@@ -203,12 +202,108 @@ namespace Azure.Communication.Identity.Tests
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(TestEnvironment.CommunicationExpiredTeamsToken);
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(TestEnvironment.CommunicationExpiredTeamsToken, CTEOptions.ClientId, CTEOptions.UserObjectId));
             }
             catch (RequestFailedException ex)
             {
                 Assert.NotNull(ex.Message);
                 Assert.True(ex.Message.Contains("401"));
+                Console.WriteLine(ex.Message);
+                return;
+            }
+            Assert.Fail("An exception should have been thrown.");
+        }
+
+        [Test]
+        [TestCase("", TestName = "GetTokenForTeamsUserWithEmptyClientIdShouldThrow")]
+        [TestCase("invalid", TestName = "GetTokenForTeamsUserWithInvalidClientIdShouldThrow")]
+        public async Task GetTokenForTeamsUserWithInvalidClientIdShouldThrow(string clientId)
+        {
+            if (TestEnvironment.ShouldIgnoreIdentityExchangeTokenTest)
+            {
+                Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
+            }
+
+            try
+            {
+                CommunicationIdentityClient client = CreateClientWithConnectionString();
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(CTEOptions.TeamsUserAadToken, clientId, CTEOptions.UserObjectId));
+            }
+            catch (RequestFailedException ex)
+            {
+                Assert.NotNull(ex.Message);
+                Assert.True(ex.Message.Contains("400"));
+                Console.WriteLine(ex.Message);
+                return;
+            }
+            Assert.Fail("An exception should have been thrown.");
+        }
+
+        [Test]
+        public async Task GetTokenForTeamsUserWithWrongClientIdShouldThrow()
+        {
+            if (TestEnvironment.ShouldIgnoreIdentityExchangeTokenTest)
+            {
+                Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
+            }
+
+            try
+            {
+                CommunicationIdentityClient client = CreateClientWithConnectionString();
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(CTEOptions.TeamsUserAadToken, CTEOptions.UserObjectId, CTEOptions.UserObjectId));
+            }
+            catch (RequestFailedException ex)
+            {
+                Assert.NotNull(ex.Message);
+                Assert.True(ex.Message.Contains("400"));
+                Console.WriteLine(ex.Message);
+                return;
+            }
+            Assert.Fail("An exception should have been thrown.");
+        }
+
+        [Test]
+        [TestCase("", TestName = "GetTokenForTeamsUserWithEmptyUserObjectIdShouldThrow")]
+        [TestCase("invalid", TestName = "GetTokenForTeamsUserWithInvalidUserObjectIdShouldThrow")]
+        public async Task GetTokenForTeamsUserWithInvalidUserObjectIdShouldThrow(string userObjectId)
+        {
+            if (TestEnvironment.ShouldIgnoreIdentityExchangeTokenTest)
+            {
+                Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
+            }
+
+            try
+            {
+                CommunicationIdentityClient client = CreateClientWithConnectionString();
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(CTEOptions.TeamsUserAadToken, CTEOptions.ClientId, userObjectId));
+            }
+            catch (RequestFailedException ex)
+            {
+                Assert.NotNull(ex.Message);
+                Assert.True(ex.Message.Contains("400"));
+                Console.WriteLine(ex.Message);
+                return;
+            }
+            Assert.Fail("An exception should have been thrown.");
+        }
+
+        [Test]
+        public async Task GetTokenForTeamsUserWithWrongUserObjectIdShouldThrow()
+        {
+            if (TestEnvironment.ShouldIgnoreIdentityExchangeTokenTest)
+            {
+                Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
+            }
+
+            try
+            {
+                CommunicationIdentityClient client = CreateClientWithConnectionString();
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(CTEOptions.TeamsUserAadToken, CTEOptions.ClientId, CTEOptions.ClientId));
+            }
+            catch (RequestFailedException ex)
+            {
+                Assert.NotNull(ex.Message);
+                Assert.True(ex.Message.Contains("400"));
                 Console.WriteLine(ex.Message);
                 return;
             }