+++ sdk/webpubsub/Azure.ResourceManager.WebPubSub/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0-beta.5 (Unreleased)
+## 1.0.0-beta.6 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,22 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.0.0-beta.5 (2022-06-28)
+
+### Breaking Changes
+
+- Prepended RP name `WebPubSub` to Private Endpoint related classes.
+- Changed `Location` property/parameter type from `string` to `AzureLocation`.
+- Changed `TenantId` property type from `string` to `Guid`.
+
+### Features Added
+
+- Added `Update` methods for Resource classes.
+
+### Bugs Fixed
+
+- Fixed serialization/deserialization bugs of `Tags` and `SystemData` properties for models inherited from ResourceData/TrackedResourceData.
 
 ## 1.0.0-beta.4 (2022-04-08)